+++ react-app/src/components/Main/ChannelBar/index.js
@@ -1,14 +1,12 @@
-
-import CreateChannelModal from "./ServerChannel";
+import AddChannelModal from "./ServerChannel";
 
 const ChannelBar = () => {
-
-
   return (
     <>
       <div className="channel-ctrl">
-          <h4 className="add-new-channel">TEXT CHANNELS</h4>
-        {<CreateChannelModal />}
+        <div>
+          {<AddChannelModal />}
+        </div>
       </div>
     </>
   );
